Make date/time work
Need to check if date is correct

+++ NcxPppp/NcxPppp.h
@@ -12,7 +12,7 @@ namespace NcxPppp
     {
     public:
 		LibRrrr();
-		String^ route(Platform::String^ path, int from, int to, double time, uint8 arriveBy);
+		String^ route(Platform::String^ path, int from, int to, uint32 time, uint8 arriveBy);
 	private:
 		std::unique_ptr<char> LibRrrr::PlatformStringToCharArray(String^ string);
 		std::wstring ToWString(const char* utf8String, unsigned int length);
